fix: Make the table of pending certificates concurrency-safe
Post #496 and in fact post the oneshot channel changes introduced in #486, the pending table of the Certificate Waiter can be accessed concurrently.

In some cases, our benchmark has revealed that the drop triggered by a call to `clear()` can interact with the `drain()` operation to create a double-send on the same oneshot::Sender, which panics.
(see comments on https://github.com/MystenLabs/narwhal/commit/1474264043110b6b02f124b915e883701908799b for evidence and re-run the bench locally at that commit to repro)

This is now visible, because the GC actually fires more often post #496.

This fixes the benchmark.